I found something called STLport for MSVC which makes MSVC 6 a bit more ansi while I was learning Ogre3D.
http://www.ogre3d.org/index.php?option=com_remository&Itemid=74&func=selectcat&cat=1

near the bottom

As such, this update addresses some of those ansi issues making the code a bit easier to run in Dev-C++

Also, I implemented a few things with shots (specifically -1 and -2 shots).

Last, I reimplemented the physics engine to use velocity based verlet.  Should allow for both hard and soft motion constraints relatively easily.

+++ Engine/Vegs.cpp
@@ -20,7 +20,7 @@ void Robot::FeedVegWaste()
 
 void Robot::FeedVegSun()
 {
-    static long counter=0;
+    static unsigned long counter=0;
     int daymod;
     float depth=0, tok=0;
 
@@ -57,7 +57,7 @@ void Robot::FeedVegSun()
             if (depth < 1)
                 depth = 1;
 
-            tok = (SimOpts.LightIntensity / pow(depth, SimOpts.Gradient) * daymod + 1);
+            tok = (SimOpts.LightIntensity / powf(depth, SimOpts.Gradient) * daymod + 1);
         }
         else
         {
